move scenarios, improve syncing
