+++ src/tests/scenarios/mergeConflictSimpleEmptyOrigin.json
