+++ src/tests/scenarios/switchVersionDoesNotDeletePersonal.json
