+++ src/tests/suites/switchVersionTests.js
@@ -4,7 +4,7 @@ import { runOperation } from "../../saga/runOperation";
 import { item, TEST_URL } from "../../constants";
 import { getItemRangeValues } from "../testHelpers";
 import { getGlobal } from "../../commands/commands";
-import * as scenarios from "../../../scenarios";
+import * as scenarios from "../scenarios";
 import { runReplaceFromBase64 } from "../../saga/create";
 import { getValues } from "../testHelpers";
